+++ src/main/java/io/nagurea/smsupsdk/campaigns/get/campaign/GetCampaignResponse.java
@@ -1,0 +1,15 @@
+package io.nagurea.smsupsdk.campaigns.get.campaign;
+
+import io.nagurea.smsupsdk.common.response.APIResponse;
+import lombok.Builder;
+import lombok.ToString;
+
+@ToString
+public class GetCampaignResponse extends APIResponse<GetCampaignResultResponse> {
+
+    @Builder
+    public GetCampaignResponse(String uid, Integer statusCode, String additionalMessage, GetCampaignResultResponse effectiveResponse) {
+        super(uid, statusCode, additionalMessage, effectiveResponse);
+    }
+
+}